requirements.txt
streamlit
requests
bs4
python-dotenv
altair<5
langchain
openai
apify-client
chromadb
google-api-python-client
tiktoken
html2text
doctran
sqlalchemy
psycopg2-binary
streamlit-aggrid
streamlit-modal